+++ catscripts/Language/Pronunciation/Pronunciation.sh
@@ -2,15 +2,30 @@
 
 KEYWORDS_PRONUNCIATION="How do you pronounce|How is .+ pronounced|pron(|o)unciation"
 
-egrep -i "$KEYWORDS_PRONUNCIATION" newpages.txt >> Pronunciation.txt
 
-PRONUNCIATION=`stat --print=%s Pronunciation.txt`
-
-if [ $PRONUNCIATION -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Pronunciation.txt"
-  export CATNAME="Pronunciation"
-  $CATEGORIZE
-fi
 
-rm Pronunciation.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Pronunciation\n"
+  fi
+
+  PRONUNCIATION=`egrep -i "$KEYWORDS_PRONUNCIATION" newpages.txt`
+
+  if [ "$PRONUNCIATION" != "" ];
+  then
+    printf "$PRONUNCIATION" > Pronunciation.txt
+    export CATFILE="Pronunciation.txt"
+    export CATNAME="Pronunciation"
+    $CATEGORIZE
+    rm Pronunciation.txt
+    unset PRONUNCIATION
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Pronunciation\n"
+  fi
+
+fi
